Move connect CLI test setup to conftest
